+++ internal/attachments/processor_purego.go
@@ -173,43 +173,3 @@ func (p *pureGoProcessor) formatSupported(format string) bool {
 	}
 	return false
 }
-
-// ThumbnailFormat picks the best output format for a thumbnail
-// derived from an input of the given format. PNG inputs stay PNG so
-// transparency survives; everything else encodes as JPEG (smaller
-// files, good enough for thumbnails). Exported so the upload pipeline
-// and tests share a single source of truth.
-func ThumbnailFormat(inputFormat string) string {
-	if inputFormat == "png" {
-		return "png"
-	}
-	return "jpeg"
-}
-
-// ThumbnailMime returns the canonical MIME type for a thumbnail
-// encoded in `format` (paired with ThumbnailFormat above).
-func ThumbnailMime(format string) string {
-	switch format {
-	case "png":
-		return "image/png"
-	case "jpeg", "jpg":
-		return "image/jpeg"
-	default:
-		return "application/octet-stream"
-	}
-}
-
-// ThumbnailExt returns the file extension for a thumbnail encoded in
-// `format`. Used to build a synthetic filename for the derived row
-// (parent's basename + variant + extension) so downloads with
-// Content-Disposition expose a sensible filename.
-func ThumbnailExt(format string) string {
-	switch format {
-	case "png":
-		return ".png"
-	case "jpeg", "jpg":
-		return ".jpg"
-	default:
-		return ".bin"
-	}
-}
